requirements.txt
boto3==1.24.39
django-cors-headers==3.13.0
django-environ==0.9.0
django-location-field==2.1.0
djangorestframework-simplejwt==5.2.0
pip-chill==1.0.1
psycopg2-binary==2.9.3
